requirements.txt
Flask==3.0.3
Pillow==10.4.0
pytesseract==0.3.10
APScheduler==3.10.4
gunicorn==23.0.0